Reduce GH API call for repo name check
Reducing API calls to decrease the chance of being rate-limited by
Github when running package analysis with Depdive.

Instead of calling the Github API twice, once to check the repo_fullname
exists and a second time to get the stats of the repo, it performs the
check (and if successful gets the stats) in a single call.

The end user will experience the same error regardless, so this is
solely a Github API call optimization.

This commit also adds an error when GITHUB_TOKEN is missing from the
environemnt.

When GITHUB_TOKEN is not in the execution environment, the command:
`depdive dep-review package-metrics .` fails with the error:
`Error: environment variable not found`

Commit adds context to the error bubbled up from `std::env::var`

Closes: #132